Fix interpolation of missing password with generated one.
When creating or updating an entry, a missing password should be replaced with a
randomly generated password. A password is considiered to be missing if
`newEntry.password` either doesn't exist, is undefined or is an empty string. To
replace a missing password underscore's extend method is used to extend a set of
defaults (including a random password) with the actual submitted data.

The problem with this approach is that underscore's extend only extends missing
properties, instead of all falsy properties.

This commit fixes this problem by deleting the password property of `newEntry`
if `newEntry.password` is falsy.