Fix: Do not save volatile property to persistence layer.
The updateUpstream method of the PasswordService class saves the entries as they
are to the persistence layer. The problem with this approach is that the
modified entries have a volatile property, which is then saved to the
persistence layer as well.

To fix this problem, this commit processes every entry and picks only the
properties which are meant to be persisted, hence dropping the undesired
volatile property.