Separate two functionalities of PasswordService.get method.
The two purposes of this function are:

 1) Getting all entries - This is achieved when calling the function with no
    arguments.
 2) Getting an entry with a specific ID - This is achieved when calling the
    function with one argument. (The ID.)

In order to be able to add more options how to get entries (Except "Return all
entries" and "Return the entry with this ID".), it is better to have a separate
function for every way of getting entries. Specifically, this commit prepares a
search functionality to be added soon.

The two separate functions are now

 1) get()
 2) getById(id)