WebCrypto: Test for specific algorithms.
When deciding whether WebCrypto is supported or not, relying solely on the
presence of the window.crypto.subtle object is not enough. Chrome on ChromeOS
and Linux implements the WebCrypto interface but does not support PBKDF2, to
name an example.

To fix this problem, this commit introduces code which checks WebCrypto support
by doing a dry run of every required algorithm. If this dry run succeeds, it is
(probably) fine to use WebCrypto instead of the JS fallback.