Terminate PBKDF2-Worker after use.
Changes made in commit 4c9dff caused the FallbackEncryptionService to actually
use the PBKDF2-Worker during tests. This proved problematic as browsers limit
the maximum number of active WebWorkers, while the test cases spawn a large
number of them.

To fix this issue, the changes in this commit terminate a PBKDF2-Worker after
it has been used.